requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
python-multipart
pillow==10.3.0
PyJWT==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy
opencv-python
numpy
httpx==0.27.2
